% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_utils.R
\name{crossBoundCumProb}
\alias{crossBoundCumProb}
\title{Estimate cumulative probabilities of crossing an efficacy or non-efficacy boundary}
\usage{
crossBoundCumProb(
  boundType = c("eff", "nonEff"),
  nAnalyses,
  monitorTrialFile,
  monitorTrialDir = NULL
)
}
\arguments{
\item{boundType}{a character string specifying if the one-sided null hypothesis is of the form \eqn{H_0: \theta \geq \theta_0} (\code{"eff"}, default) or \eqn{H_0: \theta \leq \theta_0} (\code{"nonEff"}), where \eqn{\theta} and \eqn{\theta_0} are the true hazard ratio and its value specifying the null hypothesis, respectively}

\item{nAnalyses}{a numeric value specifying the number of analyses}

\item{monitorTrialFile}{either a character string specifying an \code{.RData} file or a list outputted by \code{\link{monitorTrial}}}

\item{monitorTrialDir}{a character string specifying a path to \code{monitorTrialFile} if \code{monitorTrialFile} specifies a file name}
}
\value{
A numeric vector of estimated cumulative probabilities of crossing the specified boundary by analysis \eqn{1,\ldots,\code{nAnalyses}}.
}
\description{
Computes proportions of simulated trials that crossed either an efficacy or a non-efficacy stopping boundary by analysis \eqn{1,\ldots,\code{nAnalyses}} using an \code{.RData} output file from \code{\link{monitorTrial}}.
}
\examples{
simData <- simTrial(N=c(1000, 1000), aveVE=c(0, 0.4),
                    VEmodel="half", vePeriods=c(1, 27, 79), enrollPeriod=78,
                    enrollPartial=13, enrollPartialRelRate=0.5, dropoutRate=0.05,
                    infecRate=0.6, fuTime=156,
                    visitSchedule=c(0, (13/3)*(1:4), seq(13*6/3, 156, by=13*2/3)),
                    missVaccProb=0.05, VEcutoffWeek=26, nTrials=5,
                    stage1=78, randomSeed=300)

monitorData <- monitorTrial(dataFile=simData, stage1=78, stage2=156,
                            harmMonitorRange=c(10,75), harmMonitorAlpha=0.05,
                            effCohort=list(timingCohort=list(lagTime=0),
                                           times=c(75, 150),
                                           timeUnit="counts",
                                           lagTime=0,
                                           estimand="cox",
                                           nullVE=0,
                                           nominalAlphas=c(0.001525, 0.024501)),
                            nonEffCohorts=list(timingCohort=list(lagTime=0),
                                               times=c(75, 150),
                                               timeUnit="counts",
                                               cohort1=list(lagTime=0,
                                                            estimand="cox",
                                                            nullVE=0.4,
                                                            nominalAlphas=c(0.001525, 0.024501))),
                            lowerVEnoneff=0, highVE=1, lowerVEuncPower=0,
                            alphaHigh=0.05, alphaUncPower=0.05,
                            verbose=FALSE)

crossBoundCumProb(boundType="eff", nAnalyses=2, monitorTrialFile=monitorData)
crossBoundCumProb(boundType="nonEff", nAnalyses=2, monitorTrialFile=monitorData)

}
