% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postMonitoring.R
\name{decisionTimes}
\alias{decisionTimes}
\title{Extract the time since trial start to crossing a stopping boundary or reaching the target number of events if no stopping boundary crossed in an event-driven 2-arm trial design}
\usage{
decisionTimes(
  monitorTrialFile,
  monitorTrialDir = NULL,
  saveFile = NULL,
  saveDir = monitorTrialDir
)
}
\arguments{
\item{monitorTrialFile}{either a character vector specifying (multiple) \code{.RData} file(s) or a list of lists outputted by \code{\link{monitorTrial}}}

\item{monitorTrialDir}{a character string specifying a path to the file(s) in \code{monitorTrialFile} if \code{monitorTrialFile} specifies file name(s)}

\item{saveFile}{a character string optionally specifying an \code{.RData} file name storing the output (\code{NULL} by default)}

\item{saveDir}{a character string optionally specifying the file path for \code{saveFile} (set to \code{monitorTrialDir} by default)}
}
\value{
A list (of the same length as \code{monitorTrialFile}) of numeric vectors of times. The order of the vectors matches the order of components in \code{monitorTrialFile}. The length of each vector equals the number of simulated trials in the corresponding output list from \code{\link{monitorTrial}}.
}
\description{
Obtains times (in weeks) since trial initiation to crossing a harm, non-efficacy or efficacy boundary, or reaching the target number of events if no stopping boundary is crossed in an event-driven 2-arm trial design. The times are extracted from \code{.RData} files outputted by \code{\link{monitorTrial}}.
}
\examples{
simData <- simTrial(N=c(1000, 1000), aveVE=c(0, 0.4),
                    VEmodel="half", vePeriods=c(1, 27, 79), enrollPeriod=78,
                    enrollPartial=13, enrollPartialRelRate=0.5, dropoutRate=0.05,
                    infecRate=0.6, fuTime=156,
                    visitSchedule=c(0, (13/3)*(1:4), seq(13*6/3, 156, by=13*2/3)),
                    missVaccProb=0.05, VEcutoffWeek=26, nTrials=5,
                    stage1=78, randomSeed=300)

monitorData <- monitorTrial(dataFile=simData, stage1=78, stage2=156,
                            harmMonitorRange=c(10,75), harmMonitorAlpha=0.05,
                            effCohort=list(timingCohort=list(lagTime=0),
                                           times=c(75, 150),
                                           timeUnit="counts",
                                           lagTime=0,
                                           estimand="cox",
                                           nullVE=0,
                                           nominalAlphas=c(0.001525, 0.024501)),
                            nonEffCohorts=list(timingCohort=list(lagTime=0),
                                               times=c(75, 150),
                                               timeUnit="counts",
                                               cohort1=list(lagTime=0,
                                                            estimand="cox",
                                                            nullVE=0.4,
                                                            nominalAlphas=c(0.001525, 0.024501))),
                            lowerVEnoneff=0, highVE=1, lowerVEuncPower=0,
                            alphaHigh=0.05, alphaUncPower=0.05,
                            verbose=FALSE)

times <- decisionTimes(list(monitorData))

## alternatively, to save the .RData output file (no '<-' needed):
## decisionTimes(list(monitorData), saveFile="decisionTimes.RData")

}
