% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrial.R
\name{simTrial}
\alias{simTrial}
\title{Simulation of Multi-Arm Randomized Phase IIb/III  Efficacy Trials with Time-to-Event Endpoints}
\usage{
simTrial(
  N,
  VEmodel = c("half", "constant", "manual"),
  aveVE = NULL,
  vePeriods,
  veByPeriod = NULL,
  enrollPeriod,
  enrollPartial,
  enrollPartialRelRate,
  dropoutRate,
  infecRate,
  fuTime,
  visitSchedule,
  missVaccProb = NULL,
  VEcutoffWeek,
  nTrials,
  blockSize = NULL,
  stage1,
  saveFile = NULL,
  saveDir = NULL,
  verbose = TRUE,
  randomSeed = NULL
)
}
\arguments{
\item{N}{a numeric vector specifying the numbers of enrolled trial participants per treatment arm. The length of \code{N} equals the total number of treatment arms, and the first component of \code{N} represents the control arm.
#' @param VEmodel a character string specifying whether VE is assumed to be constant throughout the follow-up period (option \code{"constant"}), have two or three levels over time specified by \code{aveVE} (option \code{"half"}, default), or have multiple levels fully specified by \code{veByPeriod} (option \code{"manual"}). The option \code{"half"} allows either a 2- or 3-level VE model (specified by the \code{vePeriods} vector with either 2 or 3 components, respectively, and \code{aveVE}). Under the option \code{"half"}, both the 2- and 3-level VE model assumes a maximal VE in the second time interval such that, when halved in the first interval, the weighted average of VE over the first two time intervals equals \code{aveVE}. Only the first character is necessary.}

\item{aveVE}{a numeric vector containing, for each treatment arm in \code{N}, a time-averaged vaccine efficacy (VE), defined as the weighted average of VEs in two or three time intervals specified by \code{vePeriods}. \code{aveVE} and \code{vePeriods} together characterize the VE model if \code{VEmodel} is set to \code{"constant"} or \code{"half"}; otherwise, \code{aveVE} is ignored. If \eqn{\code{VEmodel} = \code{"half"}}, then \code{aveVE} is defined as follows: both the 2- and 3-level VE model assumes a maximal VE in the second time interval such that, when halved in the first interval, the weighted average of VE over the first two time intervals equals \code{aveVE}. \code{aveVE} is also applied thereafter. The components of \code{N} and \code{aveVE} correspond to each other.}

\item{vePeriods}{a numeric vector defining start times (in weeks) of time intervals with (potentially) distinct VE levels depending on the choice of \code{VEmodel}. The value 1 in the  vector signifies the beginning of follow-up. If \code{VEmodel} equals \code{"constant"}, then \code{vePeriods} must have length 1 (typically then \code{vePeriods <- 1}). If \code{VEmodel} equals \code{"half"}, then \code{vePeriods} must have length 2 or 3.}

\item{veByPeriod}{a list (\code{NULL} by default) allowing to specify the VE level for each time interval in \code{vePeriods} for each treatment arm (including the control arm). The VE model can be specified in two ways: (a) by providing a vector of 'full' VE levels for the treatment arms starting with the control arm (component \code{fullVE}) and, for each treatment arm, vectors of fractions of \code{fullVE} pertaining to the time intervals (component \code{C1} for the control arm, and \code{T1}, \code{T2}, etc. for the active arms), or (b) by providing, for each treatment arm, a vector of the actual VE levels pertaining to the time intervals (components \code{C1}, \code{T1}, \code{T2}, etc.). Note that the first component after \code{fullVE}, if it exists, must be \code{C1} for the control arm, followed by the active arms. The vector \code{fullVE}, if it exists, has the same length as the number of arms, while the vectors \code{C1}, \code{T1}, \code{T2}, etc. have the same length as the number of time intervals in \code{vePeriods}. If the VE model is specified via \code{veByPeriod}, then \code{aveVE} will be ignored.}

\item{enrollPeriod}{the final week of the enrollment period}

\item{enrollPartial}{the final week of the portion of the enrollment period with a reduced enrollment rate defined by \code{enrollPartialRelRate}}

\item{enrollPartialRelRate}{a non-negative value characterizing the fraction of the weekly enrollment rate governing enrollment from week 1 until week \code{enrollPartial}}

\item{dropoutRate}{a (prior) dropout probability within 1 year}

\item{infecRate}{a (prior) infection probability within 1 year in the control arm}

\item{fuTime}{a follow-up time (in weeks) of each participant}

\item{visitSchedule}{a numeric vector listing the visit weeks at which testing for the endpoint is conducted}

\item{missVaccProb}{a numeric vector with conditional probabilities of having missed a vaccination given the follow-up time exceeds \code{VEcutoffWeek} weeks. For each component, a separate per-protocol indicator is generated. Each per-protocol cohort includes subjects with (i) a non-missing vaccination, and (ii) follow-up time exceeding \code{VEcutoffWeek} weeks. If \code{NULL}, no per-protocol indicators are included.}

\item{VEcutoffWeek}{a time cut-off (in weeks); the follow-up time exceeding \code{VEcutoffWeek} weeks is required for inclusion in the per-protocol cohort}

\item{nTrials}{the number of trials to be simulated}

\item{blockSize}{a constant block size to be used in permuted-block randomization. The choice of \code{blockSize} requires caution to achieve the desired balance of treatment assignments within a block.}

\item{stage1}{the final week of stage 1 in a two-stage trial}

\item{saveFile}{a character string specifying the name of the output \code{.RData} file. If \code{NULL} (default), a default file name will be used.}

\item{saveDir}{a character string specifying a path for the output directory. If supplied, the output is saved as an \code{.RData} file in the directory; otherwise the output is returned as a list.}

\item{verbose}{a logical value indicating whether information on the output directory and file name should be printed out (default is \code{TRUE})}

\item{randomSeed}{sets seed of the random number generator for simulation reproducibility}
}
\value{
If \code{saveDir} is specified, the output list (named \code{trialObj}) is saved as an \code{.RData} file (the output directory path is printed); otherwise it is returned. The output object is a list with the following components:
\itemize{
  \item \code{trialData}: a list with \code{nTrials} components each of which is a \code{data.frame} with at least the variables \code{trt}, \code{entry}, \code{exit}, and \code{event} storing the treatment assignments, enrollment times, study exit times, and event indicators, respectively. The observed follow-up times can be recovered as \code{exit} - \code{entry}. Indicators of belonging to the per-protocol cohort (named \code{pp1}, \code{pp2}, etc.) are included if \code{missVaccProb} is specified.
  \item \code{NinfStage1}: a list whose components are numeric vectors with the numbers of \code{stage1} infections by treatment (\code{[1]} = control arm) for each simulated trial
  \item \code{nTrials}: the number of simulated trials
  \item \code{N}: the total number of enrolled trial participants
  \item \code{nArms}: the number of treatment arms
  \item \code{trtAssgnProbs}: a numeric vector containing the treatment assignment probabilities
  \item \code{blockSize}: the block size used for treatment assignment
  \item \code{fuTime}: the follow-up time (in weeks) of each participant
  \item \code{rates}: a list with three components: the prior weekly enrollment rate (\code{enrollment}), the prior probability of dropout within 1 week (\code{dropout}), and the prior probability of infection within 1 week (\code{infection})
  \item \code{enrollSchedule}: a \code{data.frame} summarizing information on enrollment periods and corresponding relative enrollment rates (relative to the weekly "base" enrollment rate). The column names are \code{start}, \code{end}, and \code{relativeRates}.
  \item \code{VEs}: a list with components being numeric vectors containing VE levels assumed within time periods defined by \code{vePeriods} for each active treatment arm
  \item \code{infecRates}: a \code{data.frame} summarizing information on time periods of distinct VE across all treatment arms. The variables \code{trt}, \code{start}, \code{end}, and \code{relRate} carry treatment assignment labels, first and last week of a time interval, and the pertaining assumed hazard ratio in the given interval.
  \item \code{randomSeed}: the set seed of the random number generator for simulation reproducibility
}
}
\description{
\code{simTrial} generates independent time-to-event data-sets according to a user-specified trial design. The user makes assumptions about the enrollment, dropout, and infection processes in each treatment arm.
}
\details{
All time variables use week as the unit of time. Month is defined as 52/12 weeks.

The prior weekly enrollment rate is calculated based on the duration of the enrollment periods with reduced/full enrollment rates and the total number of subjects to be enrolled.

The weekly enrollment, dropout and infection rates used for generating trial data are sampled from specified prior distributions (the prior annual dropout and infection probabilities are specified by the user). The default choice considers non-random point-mass distributions, i.e., the prior rates directly govern the accumulation of trial data.
 
Subjects' enrollment is assumed to follow a Poisson process with a time-varying rate (the argument \code{enrollPartialRelRate} characterizes a reduced enrollment rate applied to weeks 1 through \code{enrollPartial}, i.e., full enrollment starts at week \code{enrollPartial}+1). The number of enrolled subjects is determined by the vector \code{N}.
 
Dropout times are assumed to follow an exponential distribution where the probability of a dropout within 1 week is equal to \code{dropoutRate}/52.
 
Permuted-block randomization is used for assigning treatment labels. If left unspecified by the user, an appropriate block size, no smaller than 10, will computed and used.  The function \code{getBlockSize} can be used to determine appropriate block sizes (see help(getBlockSize)).
 
Infection times are generated following the VE schedule characterized by \code{aveVE}, \code{VEmodel} and \code{vePeriods}. Independent exponential times are generated within each time period of constant VE, and their minimum specifies the right-censored infection time. Exponential rates are chosen that satisfy the user-specified requirements on the treatment- and time-period-specific probabilities of an infection within 1 week (in the control arm, the infection probability within 1 week uniformly equals \code{infecRate}/52).

Infection diagnosis times are calculated according to the \code{visitSchedule}. The observed follow-up time is defined as the minumum of the infection diagnosis time, dropout time, and \code{fuTime}.
 % describe the generation of enrollment, dropout and infection rates from prior distributions to be used for data simulation (sampleRates)
 % describe the data generating process (each step), characterize the distributions/assumptions
}
\examples{
### constant VE throughout the follow-up period
simData <- simTrial(N=c(1000, rep(700, 2)), VEmodel="constant", 
                    aveVE=seq(0, 0.4, by=0.2), vePeriods=1, 
                    enrollPeriod=78, enrollPartial=13, enrollPartialRelRate=0.5, 
                    dropoutRate=0.05, infecRate=0.04, fuTime=156, 
                    visitSchedule=seq(0, 156, by=4),
                    missVaccProb=c(0,0.05,0.1,0.15), VEcutoffWeek=26, nTrials=5, 
                    stage1=78, randomSeed=300)
                    
### a 3-level VE model specified by 'aveVE' and 'vePeriods'
simData <- simTrial(N=c(1000, rep(700, 2)), VEmodel="half", 
                    aveVE=seq(0, 0.4, by=0.2), vePeriods=c(1, 27, 79), 
                    enrollPeriod=78, enrollPartial=13, enrollPartialRelRate=0.5, 
                    dropoutRate=0.05, infecRate=0.04, fuTime=156, 
                    visitSchedule=seq(0, 156, by=4),
                    missVaccProb=c(0,0.05,0.1,0.15), VEcutoffWeek=26, nTrials=5, 
                    stage1=78, randomSeed=300)
                    
### a manually entered VE model specified by 'veByPeriod' using 'fullVE'
### and fractions
simData <- simTrial(N=c(1000, rep(700, 2)), VEmodel="manual", 
                    vePeriods=c(1, 15, 27, 79),
                    veByPeriod=list(fullVE=c(0, 0.6, 0.8),
                                    C1=c(1, 1),
                                    T1=c(0.1, 2/3, 1, 0.75),
                                    T2=c(0.1, 0.5, 1, 0.9)), 
                    enrollPeriod=78, enrollPartial=13, enrollPartialRelRate=0.5, 
                    dropoutRate=0.05, infecRate=0.04, fuTime=156, 
                    visitSchedule=seq(0, 156, by=4),
                    missVaccProb=c(0,0.05,0.1,0.15), VEcutoffWeek=26, nTrials=5, 
                    stage1=78, randomSeed=300)
                    
### the same manually entered VE model as above by specifying the actual 
### VE levels in 'veByPeriod'
simData <- simTrial(N=c(1000, rep(700, 2)), VEmodel="manual", 
                    vePeriods=c(1, 15, 27, 79),
                    veByPeriod=list(C1=c(0, 0, 0, 0),
                                    T1=c(0.1, 2/3, 1, 0.75) * 0.6,
                                    T2=c(0.1, 0.5, 1, 0.9) * 0.8), 
                    enrollPeriod=78, enrollPartial=13, enrollPartialRelRate=0.5, 
                    dropoutRate=0.05, infecRate=0.04, fuTime=156, 
                    visitSchedule=seq(0, 156, by=4),
                    missVaccProb=c(0,0.05,0.1,0.15), VEcutoffWeek=26, nTrials=5, 
                    stage1=78, randomSeed=300)

### alternatively, to save the .RData output file (no '<-' needed):
###
### simTrial(N=c(1400, rep(1000, 2)), aveVE=seq(0, 0.4, by=0.2), VEmodel="half", 
###          vePeriods=c(1, 27, 79), enrollPeriod=78, enrollPartial=13, 
###          enrollPartialRelRate=0.5, dropoutRate=0.05, infecRate=0.04, fuTime=156, 
###          visitSchedule=seq(0, 156, by=4), 
###          missVaccProb=c(0,0.05,0.1,0.15), VEcutoffWeek=26, nTrials=5, 
###          stage1=78, saveDir="./", randomSeed=300)

}
\seealso{
\code{\link{monitorTrial}}, \code{\link{censTrial}}, and \code{\link{rankTrial}}
}
