% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorTrial.R
\name{monitorTrial}
\alias{monitorTrial}
\title{Group Sequential Monitoring of Simulated Efficacy Trials for the Event of Potential Harm, Non-Efficacy, and High Efficacy}
\usage{
monitorTrial(
  dataFile,
  stage1,
  stage2,
  harmMonitorRange,
  harmMonitorAlpha = 0.05,
  alphaPerTest = NULL,
  nonEffStartMethod = c("FKG", "fixed"),
  nonEffStartParams = NULL,
  nonEffIntervalUnit = c("counts", "time"),
  nonEffInterval,
  nonEffCohorts = list(times = NULL, timeUnit = "counts", timingCohort = list(lagTime =
    NULL, cohortInd = NULL), cohort1 = list(estimand = "cox", lagTime = NULL, cohortInd =
    NULL, nullVE = NULL, nominalAlphas = NULL)),
  effCohort = list(times = NULL, timeUnit = "counts", timingCohort = list(lagTime =
    NULL, cohortInd = NULL), nullVE = NULL, estimand = "cox", lagTime = NULL, cohortInd =
    NULL, nominalAlphas = NULL),
  stage1Eff = list(cohort = list(lagTime = 0, cohortInd = NULL), nullVE = NULL,
    nominalAlpha = NULL, estimand = "cox"),
  lowerVEnoneff = NULL,
  highVE,
  stage1VE,
  lowerVEuncPower = NULL,
  alphaHigh,
  alphaStage1,
  alphaUncPower = NULL,
  saveFile = NULL,
  saveDir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dataFile}{if \eqn{\code{saveDir} = \code{NULL}}, a list returned by \code{simTrial}; otherwise a name (character string) of an \code{.RData} file created by \code{simTrial}}

\item{stage1}{the final week of stage 1 in a two-stage fixed-follow-up trial}

\item{stage2}{the final week of stage 2 in a two-stage fixed-follow-up trial, i.e., the maximum total follow-up time}

\item{harmMonitorRange}{a 2-component numeric vector specifying the range of the pooled number of events (pooled over the placebo and vaccine arm accruing events the fastest) over which the type I error rate, specified in \code{harmMonitorAlpha}, shall be spent (per vaccine arm). Note that \code{harmMonitorRange} does not specify a range over which potential-harm stopping boundaries will be computed; instead, it specifies when potential-harm monitoring will start, and the range over which the type I error rate \code{harmMonitorAlpha} will be spent. If \eqn{\code{nonEffStartMethod} = \code{"FKG"}} or \code{"fixed"}, then the second value is ignored and can be replaced with \code{NA}.}

\item{harmMonitorAlpha}{a numeric value (0.05 by default) specifying the overall type I error rate for potential-harm monitoring (per vaccine arm). To turn off potential-harm monitoring, set \code{harmMonitorAlpha} equal to 0.00001.}

\item{alphaPerTest}{a per-test nominal significance level for potential-harm monitoring. If \code{NULL} (default), a per-test significance level is calculated that yields the overall type I error rate of \code{harmMonitorAlpha} at the end of \code{harmMonitorRange}.}

\item{nonEffStartMethod}{a character string specifying the method used for determining when non-efficacy monitoring is to start. The default method of Freidlin, Korn, and Gray (2010) ("\code{FKG}") calculates the minimal pooled event count (pooled over the placebo and vaccine arm accruing events the fastest) such that a hazard-ratio-based VE point estimate of 0\% would result in declaring non-efficacy, i.e., the upper bound of the two-sided \eqn{(1-\code{alphaNoneff}) 100\%} confidence interval for VE based on the asymptotic variance of the log-rank statistic equals the non-efficacy threshold specified as component \code{upperVEnonEff} in the list \code{nonEffStartParams}. If this list component is left unspecified, the argument \code{upperVEnonEff} is used as the non-efficacy threshold. The alternative method ("\code{fixed}") starts non-efficacy monitoring at a fixed pooled event count (pooled over the placebo and vaccine arm accruing events the fastest) specified by component \code{N1} in the list \code{nonEffStartParams}.}

\item{nonEffStartParams}{a list with named components specifying parameters required by \code{nonEffStartMethod} (\code{NULL} by default)}

\item{nonEffIntervalUnit}{a character string specifying whether intervals between two adjacent non-efficacy interim analyses should be event-driven (default option "\code{counts}") or calendar time-driven (option "\code{time}")}

\item{nonEffInterval}{a numeric vector (a number of events or a number of weeks) specifying the timing of non-efficacy interim analyses. If a single numeric value is specified, then all interim looks are equidistant (in terms of the number of events or weeks), and the value specifies the constant increment of information or time between two adjacent interim looks. If a numeric vector with at least two components is specified, then, following the initial interim look, the timing of subsequent interim looks is determined by (potentially differential) increments of information or time specified by this vector.}

\item{nonEffCohorts}{a named list specifying all cohorts (for both timing and analysis) used in non-efficacy monitoring. The required list components characterize the 'timing cohort,' i.e., the cohort events in which determine the analysis timepoints in an event-driven design (components \code{times}, \code{timeUnit}, and \code{timingCohort}), and the analysis cohort(s), i.e., the cohort(s) in which inference is made about the null hypothesis of non-efficacy (component \code{cohort1} is required, and \code{cohort2}, etc. are optional for additional analysis cohorts if, e.g., non-efficacy monitoring is conducted in both the ITT and per-protocol cohorts). As for the timing cohort, \code{times} specifies the analysis timepoints in terms of event counts (\eqn{\code{timeUnit} = \code{"counts"}} is the only implemented option). \code{timingCohort} is a list characterizing the timing cohort by components \code{lagTime}, a lag time (in weeks) that controls event inclusion for timing (if no lag is desired, it can be set to \code{NULL} or 0), and \code{cohortInd}, a character string naming an indicator variable included in the data frames outputted by \code{\link{simTrial}}, which also controls event inclusion for the purpose of determining analysis timepoints. The other top-level list components \code{cohort1}, \code{cohort2}, etc. are each a list that must contain a component named \code{estimand}, which can take on values \code{"cox"}, \code{"cuminc"}, or \code{"combined"}. Optional list components in \code{cohort1}, \code{cohort2}, etc. are \code{lagTime}, \code{cohortInd}, \code{nullVE}, and \code{nominalAlphas}. \code{lagTime} specifies a lag time (in weeks) that controls event inclusion for the analysis cohort. If no lag time is desired, then this component can be ignored, set to \code{NULL}, or set to 0. \code{cohortInd} is a character string naming an indicator variable included in the data frames outputted by \code{\link{simTrial}} to be used to subset participants into the analysis cohort. This component allows inclusion of, e.g., a "per-protocol" variable (e.g., by setting \code{cohortInd} to \code{"pp1"}). \code{nullVE} specifies the one-sided null hypothesis as \eqn{H_0: VE \geq \code{nullVE}}. The rejection of \eqn{H_0} constitutes evidence for non-efficacy. \code{nominalAlphas} specifies nominal significance levels in a two-arm event-driven trial design.}

\item{effCohort}{a named list specifying the timing and analysis cohorts used in group-sequential efficacy monitoring (if part of the monitoring plan). The required list components characterize the 'timing cohort,' i.e., the cohort events in which determine the analysis timepoints in an event-driven design (components \code{times}, \code{timeUnit}, and \code{timingCohort}), and a single analysis cohort, i.e., the cohort in which inference is made about the null hypothesis of efficacy (components \code{estimand}, \code{lagTime}, \code{cohortInd}, \code{nullVE}, and \code{nominalAlphas}). As for the timing cohort, \code{times} specifies the analysis timepoints in terms of event counts (\eqn{\code{timeUnit} = \code{"counts"}} is the only implemented option). \code{timingCohort} is a list characterizing the timing cohort by components \code{lagTime}, a lag time (in weeks) that controls event inclusion for timing (if no lag is desired, it can be set to \code{NULL} or 0), and \code{cohortInd}, a character string naming an indicator variable included in the data frames outputted by \code{\link{simTrial}}, which also controls event inclusion for the purpose of determining analysis timepoints. As for the analysis cohort, \code{estimand} can take on values \code{"cox"}, \code{"cuminc"}, or \code{"combined"}. \code{lagTime} specifies a lag time (in weeks) that controls event inclusion for the analysis cohort. If no lag time is desired, then this component can be ignored, set to \code{NULL}, or set to 0. \code{cohortInd} is a character string naming an indicator variable included in the data frames outputted by \code{\link{simTrial}} to be used to subset participants into the analysis cohort. This component allows inclusion of, e.g., a "per-protocol" variable (e.g., by setting \code{cohortInd} to \code{"pp1"}). \code{nullVE} specifies the one-sided primary null hypothesis as \eqn{H_0: VE \leq \code{nullVE}}. The rejection of \eqn{H_0} constitues evidence for clinically relevant efficacy. \code{nominalAlphas} specifies nominal significance levels in a two-arm event-driven trial design.}

\item{lowerVEnoneff}{specifies an additional criterion for declaring non-efficacy if the hypothesis test is based on Wald confidence interval(s). It requires that the lower bound of the two-sided Wald CI(s) for the VE estimand(s), at the confidence level determined by \code{nonEffCohorts$cohort1$nominalAlphas}, etc., lie(s) below \code{lowerVEnoneff} (typically set equal to 0). If \code{NULL} (default), this criterion is ignored.}

\item{highVE}{specifies a criterion for declaring high-efficacy: the lower bound of the two-sided \eqn{(1-\code{alphaHigh}) 100\%} confidence interval for the VE estimand lies above \code{highVE} (typically a number in the 0.5--1 range). To turn off high efficacy monitoring, set \code{highVE} equal to 1.}

\item{stage1VE}{specifies a criterion for advancement of a treatment's evaluation into Stage 2: the lower bound of the two-sided \eqn{(1-\code{alphaStage1}) 100\%} confidence interval for the VE estimand lies above \code{stage1VE} (typically set equal to 0)}

\item{lowerVEuncPower}{a numeric vector with each component specifying a one-sided null hypothesis \eqn{H_0: VE(0-\code{stage1}) \leq \code{lowerVEuncPower} \times 100\%}. Unconditional power (i.e., accounting for sequential monitoring) to reject each \eqn{H_0} is calculated, where the rejection region is defined by the lower bound of the two-sided \eqn{(1-\code{alphaUncPower}) 100\%} confidence interval for the VE estimand being above the respective component of \code{lowerVEuncPower} (typically values in the 0--0.5 range).}

\item{alphaHigh}{one minus the nominal confidence level of the two-sided confidence interval used for high efficacy monitoring}

\item{alphaStage1}{one minus the nominal confidence level of the two-sided confidence interval used for determining whether a treatment's evaluation advances into Stage 2}

\item{alphaUncPower}{one minus the nominal confidence level of the two-sided confidence interval used to test one-sided null hypotheses \eqn{H_0: VE(0-\code{stage1}) \leq \code{lowerVEuncPower} \times 100\%} against alternative hypotheses \eqn{H_1: VE(0-\code{stage1}) > \code{lowerVEuncPower} \times 100\%}. The same nominal confidence level is applied for each component of \code{lowerVEuncPower}.}

\item{saveFile}{a character string specifying the name of the output \code{.RData} file. If \code{NULL} (default), a default file name will be used.}

\item{saveDir}{a character string specifying a path for \code{dataFile}. If supplied, the output is also saved as an \code{.RData} file in this directory; otherwise the output is returned as a list.}

\item{verbose}{a logical value indicating whether information on the output directory, file name, and monitoring outcomes should be printed out (default is \code{TRUE})}
}
\value{
If \code{saveDir} (and, optionally \code{saveFile}) is specified, the output list (named \code{out}) is saved as an \code{.RData} file in \code{saveDir} (the path to \code{saveDir} is printed); otherwise it is returned. The output object is a list of length equal to the number of simulated trials, each of which is a list of length equal to the number of treatment arms, each of which is a list with (at least) the following components:
\itemize{
  \item \code{boundHit}: a character string stating the monitoring outcome in this treatment arm, i.e., one of \code{"Harm"}, \code{"NonEffInterim"}, \code{"NonEffFinal"}, \code{"Eff"}, or \code{"HighEff"}. The first four outcomes can occur in Stage 1, whereas the last outcome can combine data over Stage 1 and Stage 2.
  \item \code{stopTime}: the time of hitting a stopping boundary since the first subject enrolled in the trial
  \item \code{stopInfectCnt}: the pooled number of infections at \code{stopTime}
  \item \code{summObj}: a \code{data.frame} containing summary information from each non-/high efficacy interim analysis
  \item \code{finalHRci}: the final CI for the hazard ratio, available if \code{estimand!="cuminc"} and there is at least 1 infection in each arm
  \item \code{firstNonEffCnt}: the number of infections that triggered non-efficacy monitoring (if available)
  \item \code{totInfecCnt}: the total number of \code{stage1} (\code{stage2} if \code{boundHit = "HighEff"}) infections
  \item \code{totInfecSplit}: a table with the numbers of \code{stage1} (\code{stage2} if \code{boundHit = "HighEff"}) infections in the treatment and control arm
  \item \code{lastExitTime}: the time between the first subject's enrollment and the last subject's exiting from the trial
}
}
\description{
\code{monitorTrial} applies a group sequential monitoring procedure to data-sets generated by \code{simTrial}, which may result in modification or termination of each simulated trial.
}
\details{
All time variables use week as the unit of time. Month is defined as 52/12 weeks.

Potential harm monitoring starts at the \code{harmMonitorRange[1]}-th infection pooled over the placebo group and the vaccine regimen that accrues infections the fastest. The potential harm analyses continue at each additional infection until the first interim analysis for non-efficacy. The monitoring is implemented with exact one-sided binomial tests of H0: \eqn{p \le p0} versus H1: \eqn{p > p0}, where \eqn{p} is the probability that an infected participant was assigned to the vaccine group, and \eqn{p0} is a fixed constant that represents the null hypothesis that an infection is equally likely to be assigned vaccine or placebo. Each test is performed at the same prespecified nominal/unadjusted alpha-level (\code{alphaPerTest}), chosen based on simulations such that, for each vaccine regimen, the overall type I error rate by the \code{harmMonitorRange[2]}-th arm-pooled infection (i.e., the probability that the potential harm boundary is reached when the vaccine is actually safe, \eqn{p = p0}) equals \code{harmMonitorAlpha}.

Non-efficacy is defined as evidence that it is highly unlikely that the vaccine has a beneficial effect measured as VE(0--\code{stage1}) of \code{upperVEnoneff} x 100\% or more. The non-efficacy analyses for each vaccine regimen will start at the first infection (pooled over the vaccine and placebo arm) determined by \code{nonEffStartMethod}. Stopping for non-efficacy will lead to a reported two-sided (1-\code{alphaNoneff}) x 100\% CI for VE(0--\code{stage1}) with, optionally, the lower confidence bound below \code{lowerVEnoneff} and the upper confidence bound below \code{upperVEnoneff}, where \code{estimand} determines the choice of the VE(0--\code{stage1}) estimand. This approach is similar to the inefficacy monitoring approach of Freidlin, Korn, and Gray (2010). If \code{estimand = "combined"}, stopping for non-efficacy will lead to reported (1-\code{alphaNoneff}) x 100\% CIs for both VE parameters with, optionally, lower confidence bounds below \code{lowerVEnoneff} and upper confidence bounds below \code{upperVEnoneff}. If \code{laggedMonitoring = TRUE}, stopping for non-efficacy will lead to reported (1-\code{alphaNoneff}) x 100\% CIs for both VE(0--\code{stage1}) and VE(\code{lagTime}--\code{stage1}) with, optionally, lower confidence bounds below \code{lowerVEnoneff} and upper confidence bounds below \code{upperVEnoneff}.

High efficacy monitoring allows early detection of a highly protective vaccine if there is evidence that VE(0--\code{stage2}) \eqn{>} \code{highVE} x 100\%. It is synchronized with non-efficacy monitoring during Stage 1, and a single high-efficacy interim analysis during Stage 2 is conducted halfway between the end of Stage 1 and the end of the trial. While monitoring for potential harm and non-efficacy restricts to \code{stage1} infections, monitoring for high efficacy counts all infections during \code{stage1} or \code{stage2}, given that early stopping for high efficacy would only be warranted under evidence for durability of the efficacy.

The following principles and rules are applied in the monitoring procedure:
\itemize{
  \item Exclude all follow-up data from the analysis post-unblinding (and include all data pre-unblinding).
  \item The monitoring is based on modified ITT analysis, i.e., all subjects documented to be free of the study endpoint at baseline are included and analyzed according to the treatment assigned by randomization, ignoring how many vaccinations they received (only pre-unblinding follow-up included).
  \item If a vaccine hits the harm boundary, immediately discontinue vaccinations and accrual into this vaccine arm, and unblind this vaccine arm (continue post-unblinded follow-up until the end of Stage 1 for this vaccine arm).  
  \item If a vaccine hits the non-efficacy boundary, immediately discontinue vaccinations and accrual into this vaccine arm, keep blinded and continue follow-up until the end of Stage 1 for this vaccine arm. 
  \item If and when the last vaccine arm hits the non-efficacy (or harm) boundary, discontinue vaccinations and accrual into this vaccine arm, and unblind (the trial is over, completed in Stage 1).
  \item Stage 1 for the whole trial is over on the earliest date of the two events: (1) all vaccine arms have hit the harm or non-efficacy boundary; and (2) the last enrolled subject in the trial reaches the final \code{stage1} visit.
  \item Continue blinded follow-up until the end of Stage 2 for each vaccine arm that reaches the end of \code{stage1} with a positive efficacy (as defined by \code{stage1VE}) or high efficacy (as defined by \code{highVE}) result.
  \item If at least one vaccine arm reaches the end of \code{stage1} with a positive efficacy or high efficacy result, continue blinded follow-up in the placebo arm until the end of Stage 2.
  \item Stage 2 for the whole trial is over on the earliest date of the two events: (1) all subjects in the placebo arm and each vaccine arm that registered efficacy or high efficacy in \code{stage1} have failed or been censored; and (2) all subjects in the placebo arm and each vaccine arm that registered efficacy or high efficacy in \code{stage1} have completed the final \code{stage2} visit.
}

The above rules have the following implications:
\itemize{
  \item If a vaccine hits the non-efficacy boundary but Stage 1 for the whole trial is not over, then one includes in the analysis all follow-up through the final \code{stage1} visit for that vaccine regimen, including all individuals accrued up through the date of hitting the non-efficacy boundary (which will be the total number accrued to this vaccine arm).
  \item If a vaccine hits the harm boundary, all follow-up information through the date of hitting the harm boundary is included for this vaccine; no follow-up data are included after this date.
  \item If and when the last vaccine arm hits the non-efficacy (or harm) boundary, all follow-up information through the date of hitting the non-efficacy (or harm) boundary is included for this vaccine; no follow-up data are included after this date.
}
}
\examples{
simData <- simTrial(N=c(1000, 1000), aveVE=c(0, 0.4),
                    VEmodel="half", vePeriods=c(1, 27, 79), enrollPeriod=78,
                    enrollPartial=13, enrollPartialRelRate=0.5, dropoutRate=0.05,
                    infecRate=0.06, fuTime=156, visitSchedule=seq(0, 156, by=4),
                    missVaccProb=0.05, VEcutoffWeek=26, nTrials=5,
                    stage1=78, randomSeed=300)
                    
### trial design: fixed follow-up; cumulative incidence-based VE estimand;
                  no efficacy monitoring; non-efficacy monitoring using 
                  the Freidlin et al. method; hypothesis tests based on Wald 
                  confidence intervals
### RIGHT NOW, THE BELOW CALL IS BROKEN
monitorData <- monitorTrial(dataFile=simData, stage1=78, stage2=156,
                            harmMonitorRange=c(10, NA), harmMonitorAlpha=0.05,
                            nonEffStartMethod="FKG", nonEffInterval=20,
                            nonEffCohorts=list(timeUnit="counts",
                                               # right now 'timingCohort' is required
                                               timingCohort=list(lagTime=0),
                                               cohort1=list(estimand="cuminc",
                                                            nullVE=0.4,
                                                            nominalAlphas=0.025)),
                            # it appears that right now 'effCohort' must be specified
                            # even when there is no efficacy monitoring;
                            # this call of monitorTrial() still doesn't run
                            # because it requires event counts for timing
                            effCohort=list(timeUnit="counts",
                                           timingCohort=list(lagTime=0)),
                            lowerVEnoneff=0, highVE=0.7, lowerVEuncPower=0, 
                            alphaHigh=0.05, alphaUncPower=0.05)
                            
### trial design: event-driven; hazard-based VE estimand;
                  harmonized efficacy and non-efficacy monitoring; 
                  hypothesis tests using the score test in the Cox model
### THE SCORE TEST OPTION NEEDS TO BE ADDED
monitorData <- monitorTrial(dataFile=simData, stage1=78, stage2=156,
                            harmMonitorRange=c(10, 50), harmMonitorAlpha=0.05,
                            nonEffCohorts=list(
                              times=c(50, 100, 150),
                              timeUnit="counts",
                              timingCohort=list(lagTime=0),
                              cohort1=list(estimand="cox",
                                           nullVE=0.4,
                                           nominalAlphas=c(0.0001, 0.0060, 0.0231))),
                            effCohort=list(times=c(50, 100, 150),
                                           timeUnit="counts",
                                           timingCohort=list(lagTime=0),
                                           estimand="cox",
                                           nullVE=0,
                                           nominalAlphas=c(0.0001, 0.0060, 0.0231)),
                            highVE=1, lowerVEuncPower=0, 
                            alphaHigh=0.05, alphaUncPower=0.05)
   
### alternatively, to save the .RData output file (no '<-' needed):
###
### simTrial(N=c(1000, 1000), aveVE=c(0, 0.4),
###          VEmodel="half", vePeriods=c(1, 27, 79), enrollPeriod=78,
###          enrollPartial=13, enrollPartialRelRate=0.5, dropoutRate=0.05,
###          infecRate=0.06, fuTime=156, visitSchedule=seq(0, 156, by=4),
###          missVaccProb=0.05, VEcutoffWeek=26, nTrials=5,
###          stage1=78, saveDir="./", randomSeed=300)
###
### THIS CALL NEEDS TO BE REVISED TO A SIMPLE BUT WORKING CODE
### THE INTENT HERE IS ONLY TO ILLUSTRATE HOW OUTPUT FILES CAN BE READ IN
### THE PURPOSE IS NOT TO SHOW ANY ADDITIONAL TRIAL DESIGNS
### monitorTrial(dataFile=
###              "simTrial_nPlac=1400_nVacc=1000_1000_aveVE=0.2_0.4_infRate=0.04.RData", 
###              stage1=78, stage2=156, harmMonitorRange=c(10,100), alphaPerTest=NULL, 
###              nonEffStartMethod="FKG", nonEffInterval=20, lowerVEnoneff=0, 
###              upperVEnoneff=0.4, highVE=0.7, stage1VE=0, lowerVEuncPower=0, 
###              alphaNoneff=0.05, alphaHigh=0.05, alphaStage1=0.05, alphaUncPower=0.05, 
###              estimand="cuminc", lagTime=26, saveDir="./")

}
\references{
Freidlin B., Korn E. L., and Gray R. (2010), A general inefficacy interim monitoring rule for randomized clinical trials. \emph{Clinical Trials} 7(3):197-208.
}
\seealso{
\code{\link{simTrial}}, \code{\link{censTrial}}, \code{\link{rankTrial}}, \code{\link{estHRbound}}, \code{\link{crossBoundCumProb}}, and \code{\link{decisionTimes}}
}
